#update common data to support counties :
- add counties data and scripts
- remove census tracts scripts and data
