+++ .gitattributes
@@ -1,2 +1,1 @@
-census/*.zip filter=lfs diff=lfs merge=lfs -text
-census-2017/*.zip filter=lfs diff=lfs merge=lfs -text
+counties/*.zip filter=lfs diff=lfs merge=lfs -text
