--- data/census/.gitignore
@@ -1,3 +1,0 @@
-*.json
-
-
