--- data/census/census0.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4499cad7fcbf2265097450f5d1f5ab9d14e80bc2b10471e622314b4ba5c9f1ca
-size 49715138
