--- data/census/census1.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9ba26c517a1c1a07a412c129068ebeaa4b48f7a9f18cf3faeb70cb86f3f110e8
-size 38778078
