--- data/census/census2.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:514f46e6d59a6a950849c19e4c90ac4957e385982f4eaeb6e542ff8ef71d3997
-size 46689629
