--- data/census/census3.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1facb22c56e32aad8be41cbfcd1618501357aa5d815756cbe03dfbfa543b34a6
-size 41561147
