--- data/census/census4.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5aacea0d6e83b94706f72aa881870466e334731b956b908218019aa408fe133b
-size 62414337
