--- data/census/census5.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:22de341da29f71427c4d787df6c2ee0847060dcfd1f01ae88557cf7f183902d3
-size 27801938
