--- data/census/census6.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:fe9b6f515c9a506a392e7642ca270a85d1b1e0eafa2c4b2bf05d97fce5ea61b5
-size 602862
