--- data/census/census7.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ad0b487cc84aea39d316eef29760155faa516f062d696f64e3f8ea1b7bb7f9d4
-size 2828896
